Conflicts:
	src/mongo/db/storage/wiredtiger/SConscript

+++ src/mongo/db/storage/wiredtiger/wiredtiger_record_store.cpp
@@ -409,7 +409,7 @@ namespace mongo {
                                        ValidateResults* results,
                                        BSONObjBuilder* output ) const {
 
-        uint64_t nrecords = 0;
+        long long nrecords = 0;
         boost::scoped_ptr<RecordIterator> iter( getIterator( txn ) );
         while( !iter->isEOF() ) {
             ++nrecords;
